move static data to constants file
